+++ RELEASES.md
@@ -11,12 +11,20 @@ Bug fixes
 * Fixes a bug in `Module::set_fn_4_mut`.
 * Module API's now properly handle `&str` and `String` parameters.
 * Indexers are available under `no_object`.
+* Registered operator-assignment functions (e.g. `+=`) now work correctly.
+
+Breaking changes
+----------------
+
+* `Engine::register_set_result` and `Engine::register_indexer_set_result` now take a function that returns `Result<(), Box<EvalAltResult>>`.
+* `Engine::register_indexer_XXX` and `Module::set_indexer_XXX` panic when the type is `Arrary`, `Map` or `String`.
 
 New features
 ------------
 
 * Plugins support via procedural macros.
 * Scripted functions are allowed in packages.
+* `parse_int` and `parse_float` functions.
 
 
 Version 0.18.3